Add --numeric-io flag and experiment with new IO.
The --numeric-io flag allows for I/O to be performed directly as
integers, instead of converting back and forth from numbers/characters.

Instead of using Python's `input` builtin now we resort to
`sys.stdin.read()` to read input from the user; as of now it seems to
handle better the cases that we might be interested in. For example, now
we can pipe a whole file as input to the program. Previously, doing so
would only give as input the first line of the file.