Fix bug with input during `--numeric-io`.
When doing numeric input we need to split the input into lines and then
remove the extra final line. We were doing this in the wrong place,
which meant we would try to split the initial input in lines whenever we
read another piece of input, instead of doing it just once.