Fix bug with wrapping.
The function `_W = lambda v: (_w and v%_w) or v` was not wrapping
correctly when `_w` is different from `0` and the value `v` is exactly
`_w`. E.g. if `_w = v = 256`, then `_W(256) = (256 and 0) or 256`
evaluates to `256` when it should evaluate to `0`.
We can only allow the
right hand side of the `or` to return if `_w` is `0`.