Fix #11.
According to issue #11 brainetry programs cannot handle infinite streams
of input because they try to read _all_ the input as soon as an input
operator is encountered.
With this commit this has been fixed, as the input is now made gradually
and by making use of `sys.stdin.read(1)` and `sys.stdin.readline()` for
`--numeric-io` mode.

Because of this commit we no longer return `i`, the input left to
consume, at the end of the program.